DEV-37198; refactor: remove fixed values from options
