+++ Acrolinx.Sidebar/Util/Changetracking/DiffInputFormat.cs
@@ -1,0 +1,10 @@
+﻿/* Copyright (c) 2024 Acrolinx GmbH */
+
+namespace Acrolinx.Sdk.Sidebar.Util.Changetracking
+{
+    public enum DiffInputFormat
+    {
+        MARKUP,
+        TEXT
+    }
+}
